Solution to Bubble Sort for Strings

+++ Bubble Sort for Strings@@67c9a8aa6e98bd154cc82f40/code.c
@@ -3,9 +3,9 @@ int bubbleSort(char *arr[],int n){
     for(int i=0;i<n-1;i++){
         for(int j=0;j<n-i-1;j++){
             if(arr[j]>arr[j+1]){
-                int temp=(int)arr[j];
-                (int)arr[j]=(int)arr[j+1];
-                (int)arr[j+1]=(int)temp;
+                int temp=arr[j];
+                arr[j]=arr[j+1];
+                arr[j+1]=temp;
             }
         }
     }
